# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FarmEasy API - Indoor farm cycle management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts by status
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [current, resolved, dismissed]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a new alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlertStatus
      tags: [alerts]
      summary: Resolve or dismiss an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertStatusUpdate"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts/{id}/action:
    post:
      operationId: takeAlertAction
      tags: [alerts]
      summary: Take action on an alert (resolves it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertActionInput"
      responses:
        "200":
          description: Alert resolved with action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bad-trays/manual-checks:
    post:
      operationId: createBadTrayEntry
      tags: [bad-trays]
      summary: Create a bad tray manual check entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BadTrayEntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualCheck"
components:
  schemas:
...
    ManualCheck:
      type: object
      properties:
        id:
          type: integer
        cycleId:
          type: integer
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        isBadTrays:
          type: boolean
        issue:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, cycleId, fullTrays, halfTrays, isBadTrays, photoUrls, createdAt]
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        severity:
          type: string
          enum: [critical, warning]
        status:
          type: string
          enum: [current, resolved, dismissed]
        actionType:
          type: ["string", "null"]
        actionNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, title, severity, status, createdAt]
    AlertInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        severity:
          type: string
          enum: [critical, warning]
      required: [title, severity]
    AlertStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
      required: [status]
    AlertActionInput:
      type: object
      properties:
        actionType:
          type: string
        notes:
          type: string
      required: [actionType]
...
    BadTrayEntryInput:
      type: object
      properties:
        cycleId:
          type: integer
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        issue:
          type: string
        notes:
          type: string
      required: [cycleId, issue]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]