# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FarmEasy API - Indoor farm cycle management
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertStatus } from './alertStatus';

export interface Alert {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  severity: AlertSeverity;
  status: AlertStatus;
  /** @nullable */
  actionType?: string | null;
  /** @nullable */
  actionNotes?: string | null;
  createdAt: Date;
  /** @nullable */
  resolvedAt?: Date | null;
}